+++ test/brands/ibm_test.dart
@@ -1,0 +1,23 @@
+import 'dart:ui';
+
+import 'package:flutter_brand_palettes/flutter_brand_palettes.dart';
+import 'package:flutter_test/flutter_test.dart';
+
+/// [ibm colors](https://usbrandcolors.com/ibm-colors/)
+void main() {
+  group('Ibm Palette', () {
+    test('black', () {
+      expect(const Ibm.black().color, const Color(0xFF000000));
+    });
+    test('blue', () {
+      expect(const Ibm.blue().color, const Color(0xFF006699));
+    });
+    test('white', () {
+      expect(const Ibm.white().color, const Color(0xFFFFFFFF));
+    });
+    test('call method', () {
+      const ibm = Ibm.blue();
+      expect(ibm(), ibm.color);
+    });
+  });
+}